requirements.txt
streamlit
google-generativeai
python-dotenv
datasets
s3fs
pydantic
dataclasses-json<0.7,>=0.5.7